requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
Flask==2.0.2
Flask-Cors==3.0.10
idna==3.3
importlib-metadata==4.10.0
itsdangerous==2.0.1
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
Pygments==2.11.1
python-dotenv==0.19.2
requests==2.27.1
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.7.0
